Audit log fixes (#199)
* l10n
* bugfix: Legacy Checkbox support
* bugfix: at AuditLogView:
* bugfix: Numeric values
* bugfix: SQL Server Float is the .NET CLR double type
* New AuditLog option at FormElementOptions.cs
* Remove unused dependency from FormService.cs

+++ src/Commons/Util/EnumerableHelper.cs
@@ -66,7 +66,7 @@ public static class EnumerableHelper
             case FieldType.Time:
                 return typeof(DateTime);
             case FieldType.Float:
-                return typeof(float);
+                return typeof(double);
             case FieldType.Int:
                 return typeof(int);
             case FieldType.NText:
